+++ app/controllers/admin/users_controller.rb
@@ -88,6 +88,8 @@ class Admin::UsersController < Admin::AdminController
       @user.avatar_url = attributes.fetch(:avatar_url, nil)
     end
 
+    @user.valid_password_confirmation(attributes.fetch(:password_confirmation, ''))
+
     # This fields are optional
     @user.name = attributes.fetch(:name, nil)
     @user.last_name = attributes.fetch(:last_name, nil)
@@ -99,6 +101,7 @@ class Admin::UsersController < Admin::AdminController
 
     @user.set_fields(attributes, [:available_for_hire]) if attributes[:available_for_hire].present?
 
+    raise Sequel::ValidationFailed.new(@user.errors.full_messages.join(', ')) unless @user.errors.empty?
     @user.update_in_central
     @user.save(raise_on_failure: true)